Generate final package name directly in Electron build
Before, we were using the filename electron-builder generated, then
renaming it in the build scripts. Now electron-builder lets you give a
file name template in the config:
https://github.com/electron-userland/electron-builder/commit/22746bd229

This is probably needed for auto-update, because we weren't updating
the version info files (latest.yml etc.) to match the renamed file,
and electron-updater looks at those files.